Score Manager works, correctly saves data, map correctness checked after generation

+++ RD_Colonization/Code/Data/SinglePlayerTurnData.cs
@@ -11,7 +11,7 @@ namespace RD_Colonization.Code.Data
         public int playerId = -1;
         public List<int> cashValue = new List<int>();
         public List<int> foodValue = new List<int>();
-        public List<int> scoreValue = new List<int>();
+        public List<float> scoreValue = new List<float>();
         public List<string> eventList = new List<string>();
 
         public SinglePlayerTurnData(int playerId)
